Allow preemption of built-in functions
Previously, if the user defined a function, say strstr(), which also has
a builtin version, __builtin_strstr(), but the file contained calls only
to the builtin version, the user's version would be ignored. This is a
divergence from gcc's behaviour, where the user's function would be
called instead of the builtin (even for calls directly to the builtin).
The divergence is caused by CIL deleting the apparently unused user
function. Fix it by marking the user-defined function as referenced
whenever the builtin is marked.