Remove broken alias handling, and add a simple (passing) test case.
Previously CIL's handling of gcc's alias attribute was broken. It would
create a new function, rather than a linker alias, and would not work
for varargs functions or any kind of variable. This patch removes that
code in favour of a "do nothing" implementation which correctly passes
the simple test case provided.